src/services/analysis.ts
import { AIResponse, AnalysisResult } from '../types';

const RESPONSE_CACHE = new Map<string, AnalysisResult>();
const CACHE_TIMEOUT = 1000 * 60 * 5; // 5 minutes

export class AnalysisService {
  private static validateResponse(response: string): boolean {
    return response.length > 0 && response.length < 10000;
  }

  private static async analyzeContent(response: string): Promise<AnalysisResult> {
    // Mock analysis logic - in production, this would connect to an AI service
    const score = Math.random() * 100;
    const feedback = `Response analysis complete. Score: ${score.toFixed(2)}`;
    const suggestions = [
      'Consider adding more specific examples',
      'Try to elaborate on key concepts',
      'Include practical applications'
    ];

    return { score, feedback, suggestions };
  }

  static async analyzeResponse(aiResponse: AIResponse): Promise<AnalysisResult> {
    const cacheKey = `${aiResponse.id}-${aiResponse.timestamp}`;
    
    // Check cache first
    if (RESPONSE_CACHE.has(cacheKey)) {
      return RESPONSE_CACHE.get(cacheKey)!;
    }

    // Validate response
    if (!this.validateResponse(aiResponse.response)) {
      throw new Error('Invalid response format or length');
    }

    try {
      const result = await this.analyzeContent(aiResponse.response);
      
      // Cache the result
      RESPONSE_CACHE.set(cacheKey, result);
      setTimeout(() => RESPONSE_CACHE.delete(cacheKey), CACHE_TIMEOUT);
      
      return result;
    } catch (error) {
      throw new Error(`Analysis failed: ${error}`);
    }
  }
}